Move SlackMessage up to be injected to the SlackNotifierCommand
